refactor: card 대신 deck 을 통해 최초 draw 를 하도록 변경
