+++ src/main/java/blackjack/domain/participant/Players.java
@@ -1,8 +1,8 @@
 package blackjack.domain.participant;
 
 import blackjack.domain.card.Card;
+import blackjack.domain.card.CardResponse;
 import blackjack.domain.card.Deck;
-import blackjack.domain.card.dto.CardResponse;
 import blackjack.domain.participant.exception.PlayerNotFoundException;
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -46,9 +46,7 @@ class Players {
 
     void distributeInitialCards(final Deck deck) {
         for (final Player player : players) {
-            final Card firstCard = deck.popCard();
-            final Card secondCard = deck.popCard();
-            player.drawInitialCard(firstCard, secondCard);
+            player.drawInitialCard(deck);
         }
     }
